requirements.txt
flask==2.3.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
flask-sqlalchemy==3.1.1
Flask-Migrate==4.0.5
marshmallow==3.20.1
pytest==7.4.3
pytest-flask==1.3.0